cms-admin: Mock ActionLogDialog story queries with MSW
Drop the inline Apollo mock client and manual provider from the story.
Register handlers for `ActionLogDialogGrid`, `ActionLogDialogShowVersion` and
`ActionLogDialogCompare` in `.storybook/mocks/actionLogDialogHandler.ts` so
the story uses the existing global Apollo + MSW infrastructure (same pattern
as `currentUserHandler`).
