+++ packages/admin/cms-admin/.storybook/mocks/handlers.ts
@@ -1,3 +1,4 @@
+import { actionLogDialogHandlers } from "./actionLogDialogHandler";
 import { currentUserHandler } from "./currentUserHandler";
 
-export const handlers = [currentUserHandler];
+export const handlers = [currentUserHandler, ...actionLogDialogHandlers];
